change model float64 to float32

+++ src/datas/__init__.py
@@ -2,7 +2,7 @@ import torch
 
 __all__ = ["federated_dataset"]
 
-FEATURE_TYPE = torch.float64
+FEATURE_TYPE = torch.float32
 TARGET_TYPE = torch.int16
-TARGET_TYPE_SYN = torch.float64  # torch.float64  # torch.int16
-VALUE_TYPE = torch.float64
+TARGET_TYPE_SYN = torch.float32  # torch.float64  # torch.int16
+VALUE_TYPE = torch.float32
